Extract triple dot menu into reusable TripleDotMenu component * Support horizontal and vertical orientations via makeVertical prop
